refactored code for the dashboard
