+++ frontend/src/App.css
@@ -1,4 +1,4 @@
-#root {
+/* #root {
   max-width: 1280px;
   margin: 0 auto;
   padding: 2rem;
@@ -39,4 +39,4 @@
 
 .read-the-docs {
   color: #888;
-}
+} */
