+++ frontend/src/components/SidebarItem.tsx
@@ -1,0 +1,8 @@
+interface SidebarItemProps {
+
+}
+export const SidebarItem = () => (
+    <div>
+        
+    </div>
+)
