+++ frontend/src/components/SidebarLogo.tsx
@@ -1,0 +1,4 @@
+import SidebarLogoSVG from "../assets/SidebarLogo.svg"
+export const SidebarLogo = () => (
+    <img src={SidebarLogoSVG} className="h-16"/>
+)
