--- frontend/src/pages/home/Dashboard.tsx
@@ -1,11 +1,0 @@
-//This main page is used to test on implementing main page including dashboard, team, project, etc
-
-export default function Dashboard() {
-  return (
-    <div className='bg-white rounded-lg pb-4 shadow h-[200vh]'>
-      <h1 className=''>Hello Dashboard</h1>
-    </div>
-  );
-};
-
-
