+++ frontend/tailwind.config.js
@@ -2,7 +2,11 @@
 export default {
   content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
   theme: {
-    extend: {},
+    extend: {
+      colors: {
+        'primary-red': '#b64859'
+      }
+    },
   },
   plugins: [],
 };
